% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supporting_functions.R
\name{b.est.clarabel.ols}
\alias{b.est.clarabel.ols}
\title{Clarabel W Optimization - OLS (Unconstrained)}
\usage{
b.est.clarabel.ols(A, Z, C = NULL, J, KM, w.constr, V)
}
\arguments{
\item{A}{Treated unit features matrix}

\item{Z}{Donor features matrix}

\item{C}{Optional constant-term matrix.}

\item{J}{Number of donor units}

\item{KM}{Number of additional variables}

\item{w.constr}{Constraint specification}

\item{V}{Feature weighting matrix}
}
\value{
Numeric vector of weights
}
\description{
Solve unconstrained quadratic programming problem using clarabel
}
