% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supporting_functions.R
\name{b.est.clarabel.ridge}
\alias{b.est.clarabel.ridge}
\title{Clarabel W Optimization - Ridge}
\usage{
b.est.clarabel.ridge(A, Z, J, KM, w.constr, V)
}
\arguments{
\item{A}{Treated unit features matrix}

\item{Z}{Donor features matrix}

\item{J}{Number of donor units}

\item{KM}{Number of additional variables}

\item{w.constr}{Constraint specification}

\item{V}{Feature weighting matrix}
}
\value{
Numeric vector of weights
}
\description{
Solve ridge-constrained QP using clarabel (L2 norm constraint)
}
