% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supporting_functions.R
\name{b.est.clarabel.simplex}
\alias{b.est.clarabel.simplex}
\title{Clarabel W Optimization - Simplex}
\usage{
b.est.clarabel.simplex(A, Z, C = NULL, J, KM, w.constr, V)
}
\arguments{
\item{A}{Treated unit features matrix}

\item{Z}{Donor features matrix}

\item{C}{Optional constant-term matrix.}

\item{J}{Number of donor units}

\item{KM}{Number of additional variables}

\item{w.constr}{Constraint specification}

\item{V}{Feature weighting matrix}
}
\value{
Numeric vector of weights
}
\description{
Solve simplex-constrained QP using clarabel (sum to 1, non-negative)
}
