% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supporting_functions.R
\name{b.est.clarabel.unified}
\alias{b.est.clarabel.unified}
\title{Unified Clarabel W Optimization Dispatcher}
\usage{
b.est.clarabel.unified(A, Z, C = NULL, J, KM, w.constr, V)
}
\arguments{
\item{A}{Treated unit features matrix}

\item{Z}{Donor features matrix}

\item{C}{Optional constant-term matrix.}

\item{J}{Number of donor units}

\item{KM}{Number of additional variables}

\item{w.constr}{Constraint specification}

\item{V}{Feature weighting matrix}
}
\value{
Numeric vector of weights
}
\description{
Dispatch to appropriate clarabel constraint-specific function
}
