% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supporting_functions.R
\name{b.est.cvxr.fallback}
\alias{b.est.cvxr.fallback}
\title{CVXR fallback function for W optimization}
\usage{
b.est.cvxr.fallback(A, Z, C = NULL, J, KM, w.constr, V, CVXR.solver)
}
\arguments{
\item{A}{Treated unit features matrix}

\item{Z}{Donor features matrix}

\item{C}{Optional constant-term matrix.}

\item{J}{Number of donor units}

\item{KM}{Number of additional variables}

\item{w.constr}{Constraint specification}

\item{V}{Feature weighting matrix}

\item{CVXR.solver}{CVXR solver to use}
}
\value{
Numeric vector of weights
}
\description{
Solve W optimization using CVXR (original implementation)
}
