% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_bootstrap.R
\name{calculate_bootstrap_pvalues}
\alias{calculate_bootstrap_pvalues}
\title{Calculate Bootstrap P-values}
\usage{
calculate_bootstrap_pvalues(bootstrap_effects, actual_effects, verbose = FALSE)
}
\arguments{
\item{bootstrap_effects}{List of bootstrap effect distributions}

\item{actual_effects}{List of actual treatment effects}

\item{verbose}{Logical for verbose output}
}
\value{
Data.table with p-values
}
\description{
Calculate Bootstrap P-values
}
