% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_advanced.R
\name{calculate_shapviz_interactions}
\alias{calculate_shapviz_interactions}
\title{Calculate SHAP Values and Interactions using shapviz}
\usage{
calculate_shapviz_interactions(shap_results, treated_unit, top_n)
}
\arguments{
\item{shap_results}{List. Results from XGBoost analysis.}

\item{treated_unit}{Character. Name of treated unit.}

\item{top_n}{Integer. Number of top interactions to return.}
}
\value{
List with shapviz SHAP results.
}
\description{
Calculate SHAP Values and Interactions using shapviz
}
\keyword{internal}
