% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supporting_functions.R
\name{collect.optimx}
\alias{collect.optimx}
\title{Collect and Process optimx Results}
\usage{
collect.optimx(res, opt = "min")
}
\arguments{
\item{res}{Results object from optimx optimization.}

\item{opt}{Character. Direction of optimization: "min" for minimum, "max" for maximum.}
}
\value{
Processed optimization results.
}
\description{
Helper function to process and reorder optimization results from optimx.
}
