% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supporting_functions.R
\name{collect_solver_info}
\alias{collect_solver_info}
\title{Collect solver information from specification curve results}
\usage{
collect_solver_info(spec_results)
}
\arguments{
\item{spec_results}{List of scest results (e.g., from specification curve analysis)}
}
\value{
Data frame with solver information for each specification
}
\description{
Collect and summarize solver information from multiple SCM results
}
