% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_bootstrap.R
\name{create_bootstrap_dataset}
\alias{create_bootstrap_dataset}
\title{Create Bootstrap Dataset by Unit-level Resampling}
\usage{
create_bootstrap_dataset(
  dataset_null,
  bootstrap_units,
  col_name_unit,
  treated_unit
)
}
\arguments{
\item{dataset_null}{Original dataset with null enforced}

\item{bootstrap_units}{Bootstrapped unit names}

\item{col_name_unit}{Column name for unit identifier}

\item{treated_unit}{Name of treated unit}
}
\value{
Bootstrap dataset
}
\description{
Create Bootstrap Dataset by Unit-level Resampling
}
