% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_bootstrap.R
\name{enforce_null_hypothesis}
\alias{enforce_null_hypothesis}
\title{Enforce Null Hypothesis by Subtracting Treatment Effects}
\usage{
enforce_null_hypothesis(dataset, sc.pred, actual_effects, verbose = FALSE)
}
\arguments{
\item{dataset}{Original dataset}

\item{sc.pred}{SC estimation results}

\item{actual_effects}{List of actual treatment effects by outcome model}

\item{verbose}{Logical for verbose output}
}
\value{
Dataset with null hypothesis enforced
}
\description{
Enforce Null Hypothesis by Subtracting Treatment Effects
}
