% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_processing.R
\name{generate_var_name}
\alias{generate_var_name}
\title{Generate Variable Name}
\usage{
generate_var_name(cov_spec, cov_name, period = NULL)
}
\description{
Generate Variable Name
}
\keyword{internal}
