% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_utils.R
\name{get_optimal_solver}
\alias{get_optimal_solver}
\title{Get Optimal Solver for Problem Characteristics}
\usage{
get_optimal_solver(
  n_vars,
  constraint_type,
  available_solvers = CVXR::installed_solvers()
)
}
\arguments{
\item{n_vars}{Number of variables}

\item{constraint_type}{Type of constraint ("simplex", "lasso", "ridge", "ols")}

\item{available_solvers}{Character vector of available solvers}
}
\value{
Character, recommended solver name
}
\description{
Selects the most efficient CVXR solver based on problem size and constraints.
}
\keyword{internal}
