% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supporting_functions.R
\name{get_solver_info}
\alias{get_solver_info}
\title{Extract solver information from SCM results}
\usage{
get_solver_info(scm_result)
}
\arguments{
\item{scm_result}{Result from scest() function}
}
\value{
List containing solver information
}
\description{
Extract information about which solvers were used for W and V optimization
}
