% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_utils.R
\name{MatrixCache}
\alias{MatrixCache}
\title{Cached Matrix Computation Manager}
\description{
Manages caching of expensive matrix computations to avoid redundant calculations
in specification curve analysis where many similar matrices are processed.
}
\keyword{internal}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{cache_store}}{Internal hash table for storing computed matrices}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MatrixCache-new}{\code{MatrixCache$new()}}
\item \href{#method-MatrixCache-get_cache_key}{\code{MatrixCache$get_cache_key()}}
\item \href{#method-MatrixCache-store_result}{\code{MatrixCache$store_result()}}
\item \href{#method-MatrixCache-get_result}{\code{MatrixCache$get_result()}}
\item \href{#method-MatrixCache-has_result}{\code{MatrixCache$has_result()}}
\item \href{#method-MatrixCache-clear_cache}{\code{MatrixCache$clear_cache()}}
\item \href{#method-MatrixCache-get_stats}{\code{MatrixCache$get_stats()}}
\item \href{#method-MatrixCache-clone}{\code{MatrixCache$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MatrixCache-new"></a>}}
\if{latex}{\out{\hypertarget{method-MatrixCache-new}{}}}
\subsection{Method \code{new()}}{
Initialize the cache
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MatrixCache$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MatrixCache-get_cache_key"></a>}}
\if{latex}{\out{\hypertarget{method-MatrixCache-get_cache_key}{}}}
\subsection{Method \code{get_cache_key()}}{
Generate cache key from matrix characteristics
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MatrixCache$get_cache_key(matrix, operation)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{matrix}}{Matrix to generate key for}

\item{\code{operation}}{Character string describing the operation}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MatrixCache-store_result"></a>}}
\if{latex}{\out{\hypertarget{method-MatrixCache-store_result}{}}}
\subsection{Method \code{store_result()}}{
Store computation result in cache
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MatrixCache$store_result(key, result)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{Cache key}

\item{\code{result}}{Computation result to store}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MatrixCache-get_result"></a>}}
\if{latex}{\out{\hypertarget{method-MatrixCache-get_result}{}}}
\subsection{Method \code{get_result()}}{
Retrieve cached result
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MatrixCache$get_result(key)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{Cache key}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MatrixCache-has_result"></a>}}
\if{latex}{\out{\hypertarget{method-MatrixCache-has_result}{}}}
\subsection{Method \code{has_result()}}{
Check if result exists in cache
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MatrixCache$has_result(key)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{Cache key}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MatrixCache-clear_cache"></a>}}
\if{latex}{\out{\hypertarget{method-MatrixCache-clear_cache}{}}}
\subsection{Method \code{clear_cache()}}{
Clear cache to free memory
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MatrixCache$clear_cache()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MatrixCache-get_stats"></a>}}
\if{latex}{\out{\hypertarget{method-MatrixCache-get_stats}{}}}
\subsection{Method \code{get_stats()}}{
Get cache statistics
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MatrixCache$get_stats()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MatrixCache-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MatrixCache-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MatrixCache$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
