% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_bootstrap.R
\name{perform_bootstrap_inference}
\alias{perform_bootstrap_inference}
\title{Perform Bootstrap Inference}
\usage{
perform_bootstrap_inference(
  dataset_null,
  sc.pred,
  actual_effects,
  n_bootstrap,
  cores,
  verbose
)
}
\arguments{
\item{dataset_null}{Dataset with null hypothesis enforced}

\item{sc.pred}{Original SC estimation results}

\item{actual_effects}{List of actual treatment effects}

\item{n_bootstrap}{Number of bootstrap replications}

\item{cores}{Number of cores for parallel processing}

\item{verbose}{Logical for verbose output}
}
\value{
List with bootstrap effects and estimates
}
\description{
Perform Bootstrap Inference
}
