% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_interface.R
\name{print.scest_formula}
\alias{print.scest_formula}
\title{Print Method for Formula-Based Synthetic Control}
\usage{
\method{print}{scest_formula}(x, ...)
}
\arguments{
\item{x}{An scest_formula object}

\item{...}{Additional arguments}
}
\description{
Print method for scest_formula objects
}
