% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supporting_functions.R
\name{print_solver_info}
\alias{print_solver_info}
\title{Print solver information summary}
\usage{
print_solver_info(solver_info)
}
\arguments{
\item{solver_info}{Result from get_solver_info()}
}
\description{
Print a human-readable summary of solver information
}
