% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_processing.R
\name{process_aggregate_arbitrary}
\alias{process_aggregate_arbitrary}
\title{\strong{CORE PROCESSING FUNCTIONS}
Process Pattern 1: Aggregate Over Arbitrary Periods}
\usage{
process_aggregate_arbitrary(
  dt_var,
  cov_spec,
  period.pre,
  id.var,
  time.var,
  var_name
)
}
\description{
\strong{CORE PROCESSING FUNCTIONS}
Process Pattern 1: Aggregate Over Arbitrary Periods
}
\keyword{internal}
