% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_processing.R
\name{process_aggregate_range}
\alias{process_aggregate_range}
\title{Process Pattern 2: Aggregate Over Range (Last N / First M)}
\usage{
process_aggregate_range(
  dt_var,
  cov_spec,
  period.pre,
  id.var,
  time.var,
  var_name
)
}
\description{
Process Pattern 2: Aggregate Over Range (Last N / First M)
}
\keyword{internal}
