% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_processing.R
\name{process_covariates}
\alias{process_covariates}
\title{Process Covariate Aggregations for Synthetic Control}
\usage{
process_covariates(data, covagg, period.pre, id.var, time.var)
}
\arguments{
\item{data}{Data frame or data.table in long format}

\item{covagg}{List of covariate aggregation specifications}

\item{period.pre}{Vector of pre-treatment periods}

\item{id.var}{Character name of ID variable}

\item{time.var}{Character name of time variable}
}
\value{
List of processed covariate matrices
}
\description{
Process Covariate Aggregations for Synthetic Control
}
\keyword{internal}
