% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_processing.R
\name{process_each_arbitrary}
\alias{process_each_arbitrary}
\title{Process Pattern 3: Each Period for Specified Periods}
\usage{
process_each_arbitrary(
  dt_var,
  cov_spec,
  period.pre,
  id.var,
  time.var,
  var_name
)
}
\description{
Process Pattern 3: Each Period for Specified Periods
}
\keyword{internal}
