% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_processing.R
\name{process_each_period_dt}
\alias{process_each_period_dt}
\title{Process Each Period as Separate Matching Variable}
\usage{
process_each_period_dt(
  dt_var,
  cov_name,
  period.pre,
  id.var,
  time.var,
  var_name
)
}
\description{
Process Each Period as Separate Matching Variable
}
\keyword{internal}
