% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_processing.R
\name{process_every_n_periods_dt}
\alias{process_every_n_periods_dt}
\title{Process Every N Periods}
\usage{
process_every_n_periods_dt(dt_var, n, period.pre, id.var, time.var, var_name)
}
\description{
Process Every N Periods
}
\keyword{internal}
