% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_processing.R
\name{process_first_n_dt}
\alias{process_first_n_dt}
\title{Process First N Periods}
\usage{
process_first_n_dt(dt_var, n, period.pre, id.var, time.var, var_name)
}
\description{
Process First N Periods
}
\keyword{internal}
