% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_processing.R
\name{process_growth_dt}
\alias{process_growth_dt}
\title{Process Growth Rates}
\usage{
process_growth_dt(dt_var, growth_type, id.var, time.var, var_name)
}
\description{
Process Growth Rates
}
\keyword{internal}
