% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_processing.R
\name{process_ranges_dt}
\alias{process_ranges_dt}
\title{Process Custom Ranges}
\usage{
process_ranges_dt(dt_var, ranges, cov_name, id.var, time.var, var_name)
}
\description{
Process Custom Ranges
}
\keyword{internal}
