% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_processing.R
\name{process_rolling_dt}
\alias{process_rolling_dt}
\title{Process Rolling Windows}
\usage{
process_rolling_dt(dt_var, window_size, cov_name, id.var, time.var, var_name)
}
\description{
Process Rolling Windows
}
\keyword{internal}
