% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_processing.R
\name{process_specific_periods_dt}
\alias{process_specific_periods_dt}
\title{Process Specific Periods}
\usage{
process_specific_periods_dt(dt_var, periods, id.var, time.var, var_name)
}
\description{
Process Specific Periods
}
\keyword{internal}
