% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_processing.R
\name{process_volatility_dt}
\alias{process_volatility_dt}
\title{Process Volatility Measures}
\usage{
process_volatility_dt(dt_var, vol_type, id.var, time.var, var_name)
}
\description{
Process Volatility Measures
}
\keyword{internal}
