% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_processing.R
\name{resolve_target_periods}
\alias{resolve_target_periods}
\title{\strong{HELPER FUNCTIONS}
Resolve Target Periods from Specification}
\usage{
resolve_target_periods(cov_spec, period.pre)
}
\description{
\strong{HELPER FUNCTIONS}
Resolve Target Periods from Specification
}
\keyword{internal}
