% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.V.R
\name{solve_qp_clarabel}
\alias{solve_qp_clarabel}
\title{Solve QP using Clarabel for V optimization}
\usage{
solve_qp_clarabel(c, H, A, b, l, u)
}
\arguments{
\item{c}{Linear objective coefficients}

\item{H}{Quadratic objective matrix (Hessian)}

\item{A}{Equality constraint matrix}

\item{b}{Equality constraint RHS}

\item{l}{Lower bounds}

\item{u}{Upper bounds}
}
\value{
Solution vector as matrix
}
\description{
Solve quadratic programming problem using clarabel solver
}
