% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_curve_methods.R
\name{spec_curve-methods
Print Method for Specification Curve Results}
\alias{spec_curve-methods
Print Method for Specification Curve Results}
\alias{print.spec_curve}
\title{S3 Methods for Specification Curve Objects}
\usage{
\method{print}{spec_curve}(x, ...)
}
\arguments{
\item{x}{An object of class "spec_curve" from \code{\link{spec_curve}}}

\item{...}{Additional arguments (currently unused)}
}
\value{
No return value, called for side effects (printing)
}
\description{
This file contains S3 methods for objects of class "spec_curve" created by
the \code{\link{spec_curve}} function. These methods provide a professional
interface for examining and summarizing specification curve analysis results.

Print method for specification curve objects that provides a concise
overview of the analysis results including number of specifications, treatment
effect distribution, and available inference results.
}
\details{
The print method displays:
\itemize{
\item Number of specifications estimated
\item Number of outcomes analyzed
\item Treatment effect summary statistics
\item Available inference methods and significance rates
\item Specification dimension summary
}
}
\examples{
\dontrun{
# After running specification curve analysis
spec_results <- spec_curve(...)
print(spec_results)  # or just: spec_results
}

}
