% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_interface.R
\name{standardize_constraint_spec}
\alias{standardize_constraint_spec}
\title{Standardize Constraint Specification}
\usage{
standardize_constraint_spec(constraint)
}
\arguments{
\item{constraint}{Character or list constraint specification}
}
\value{
List in scest w.constr format
}
\description{
Convert character constraint to scest format
}
\keyword{internal}
