% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_interface.R
\name{summary.scest_formula}
\alias{summary.scest_formula}
\title{Summary Method for Formula-Based Synthetic Control}
\usage{
\method{summary}{scest_formula}(object, ...)
}
\arguments{
\item{object}{An scest_formula object}

\item{...}{Additional arguments}
}
\description{
Summary method for scest_formula objects
}
