% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_processing.R
\name{validate_agg_function}
\alias{validate_agg_function}
\title{Validate Aggregation Function}
\usage{
validate_agg_function(agg_fun, cov_name)
}
\description{
Validate Aggregation Function
}
\keyword{internal}
