% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_utils.R
\name{validate_column_numeric}
\alias{validate_column_numeric}
\title{Validate Column is Numeric}
\usage{
validate_column_numeric(data, col_name, arg_name)
}
\arguments{
\item{data}{Data frame}

\item{col_name}{Column name to check}

\item{arg_name}{Name of argument (for error messages)}
}
\value{
NULL (throws error if validation fails)
}
\description{
Validate Column is Numeric
}
\keyword{internal}
