% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_processing.R
\name{validate_covariate_spec}
\alias{validate_covariate_spec}
\title{\strong{VALIDATION FUNCTIONS}
Validate Covariate Specification}
\usage{
validate_covariate_spec(cov_spec, cov_name, period.pre)
}
\description{
\strong{VALIDATION FUNCTIONS}
Validate Covariate Specification
}
\keyword{internal}
