% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_utils.R
\name{validate_list}
\alias{validate_list}
\title{Validate List Input}
\usage{
validate_list(x, arg_name, required_names = NULL, allow_null = TRUE)
}
\arguments{
\item{x}{Object to validate}

\item{arg_name}{Character. Name of argument}

\item{required_names}{Character vector of required list element names}

\item{allow_null}{Logical. Whether NULL is acceptable}
}
\value{
NULL (throws error if validation fails)
}
\description{
Validate List Input
}
\keyword{internal}
