% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_utils.R
\name{validate_logical}
\alias{validate_logical}
\title{Validate Logical Input}
\usage{
validate_logical(x, arg_name, length = 1)
}
\arguments{
\item{x}{Object to validate}

\item{arg_name}{Character. Name of argument}

\item{length}{Expected length (default = 1)}
}
\value{
NULL (throws error if validation fails)
}
\description{
Validate Logical Input
}
\keyword{internal}
