% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_processing.R
\name{validate_range_size}
\alias{validate_range_size}
\title{Validate Range Size}
\usage{
validate_range_size(range_size, total_periods, cov_name, range_type)
}
\description{
Validate Range Size
}
\keyword{internal}
