% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_utils.R
\name{validate_s3_class}
\alias{validate_s3_class}
\title{Validate S3 Object Class}
\usage{
validate_s3_class(obj, required_class, arg_name)
}
\arguments{
\item{obj}{Object to validate}

\item{required_class}{Character. Required class name}

\item{arg_name}{Name of argument (for error messages)}
}
\value{
NULL (throws error if validation fails)
}
\description{
Validate S3 Object Class
}
\keyword{internal}
