% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_utils.R
\name{validate_values_in_column}
\alias{validate_values_in_column}
\title{Validate Values Exist in Column}
\usage{
validate_values_in_column(data, col_name, values, arg_name)
}
\arguments{
\item{data}{Data frame}

\item{col_name}{Column name}

\item{values}{Values to check for}

\item{arg_name}{Name of argument (for error messages)}
}
\value{
NULL (throws error if validation fails)
}
\description{
Validate Values Exist in Column
}
\keyword{internal}
