% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_curve_methods.R
\name{plot.spec_curve}
\alias{plot.spec_curve}
\title{Plot Method for Specification Curve Results}
\usage{
\method{plot}{spec_curve}(x, ...)
}
\arguments{
\item{x}{An object of class "spec_curve" from \code{\link{spec_curve}}}

\item{...}{Additional arguments passed to \code{\link{plot_spec_curve}}}
}
\value{
A ggplot object containing the specification curve plot
}
\description{
Plot method that creates a specification curve visualization by calling
the main plotting function with appropriate defaults.
}
\details{
This is a convenience method that calls \code{\link{plot_spec_curve}} with the
specification curve data. All arguments supported by \code{plot_spec_curve}
can be passed through the \code{...} parameter.
}
\examples{
\dontrun{
spec_results <- spec_curve(...)
plot(spec_results)

# With additional options
plot(spec_results, show_shap = TRUE, show_pvalues = TRUE)
}

}
\seealso{
\code{\link{plot_spec_curve}} for detailed plotting options
}
