\name{SCMS-package}
\alias{SCMS-package}
\alias{SCMS}
\docType{package}
\title{
\packageTitle{SCMS}
}
\description{
\packageDescription{SCMS}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SCMS}
\packageIndices{SCMS}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{SCMS}

Maintainer: \packageMaintainer{SCMS}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
