% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCMs.R
\docType{package}
\name{SCMs-package}
\alias{SCMs}
\alias{SCMs-package}
\title{SCMS: Synthetic Control Methods with Specification Curve Analysis}
\description{
This package provides tools for implementing Synthetic Control Methods (SCM),
performing inference, conducting specification curve analysis, and visualizing results.
}
\author{
\strong{Maintainer}: 

Authors:
\itemize{
  \item Zubin Jelveh
}

}
\keyword{internal}
