% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scplot.R
\name{scplot}
\alias{scplot}
\title{Plot Synthetic Control Point Estimates and Prediction Interval}
\usage{
scplot(
  result,
  fig.path = NULL,
  fig.name = NULL,
  fig.format = "png",
  col.treated = "black",
  col.synth = "mediumblue",
  label.xy = NULL,
  plot.range = NULL,
  x.ticks = NULL,
  event.label = NULL,
  plot.specs = NULL,
  save.data = NULL
)
}
\arguments{
\item{result}{A class \code{"scest"} object from \code{\link{scest}}.}

\item{fig.path}{a string indicating the path where the plot(s) should be saved.}

\item{fig.name}{a string indicating the name of the plot(s). If multiple plots will be saved the command automatically
generates a numeric suffix to avoid overwriting them.}

\item{fig.format}{a string indicating the format in which the plot(s) should be saved.}

\item{col.treated}{a string specifying the color for the treated unit series. Find the full list at \url{http://sape.inf.usi.ch/quick-reference/ggplot2/colour}.}

\item{col.synth}{a string specifying the color for the synthetic unit series. Find the full list at \url{http://sape.inf.usi.ch/quick-reference/ggplot2/colour}.}

\item{label.xy}{a character list with two elements indicating the name of the axes
(eg. label.xy = list(x.lab = "Year", y.lab = "GDP growth (\%)")).}

\item{plot.range}{a numeric array indicating the time range of the plot(s).}

\item{x.ticks}{a numeric list containing the location of the ticks on the x axis.}

\item{event.label}{a list containing a character object ('lab') indicating the label of the event and
a numeric object indicating the height of the label in the plot.}

\item{plot.specs}{a list containing some specifics to be passed to ggsave (eg. img.width, img.height, dpi)}

\item{save.data}{a character specifying the name and the path of the saved dataframe containing the processed data used
to produce the plot.}
}
\value{
\item{plots}{a list containing standard ggplot object(s) that can be used for further customization.}
}
\description{
Plot treated and synthetic time series from a fitted
\code{\link{scest}} object.

Companion \href{https://www.stata.com/}{Stata} and \href{https://www.python.org/}{Python} packages are described in \href{https://arxiv.org/abs/2202.05984}{Cattaneo, Feng, Palomba, and Titiunik (2022)}.

Companion commands are \link{scdata} for data preparation, \link{scest} for
point estimation, and \link{inference_sc} for inference procedures.

For an introduction to synthetic control methods, see \href{https://www.aeaweb.org/articles?id=10.1257/jel.20191450}{Abadie (2021)} and references therein.
}
\examples{
\dontrun{
# Create example data
df <- data.frame(
  unit = rep(c("treated", "control1", "control2"), each = 20),
  time = rep(1:20, 3),
  outcome = c(rnorm(20, mean = 1:20), rnorm(20, mean = 1:20), rnorm(20, mean = 1:20))
)

# Prepare data for SCM
scm_data <- scdata(df = df, id.var = "unit", time.var = "time",
                   outcome.var = "outcome", period.pre = 1:10,
                   period.post = 11:20, unit.tr = "treated",
                   unit.co = c("control1", "control2"))

# Estimate synthetic control
result <- scest(scm_data, w.constr = list(name = "simplex"))

# Plot results
scplot(result)
}

}
\references{
\itemize{
\item{\href{https://www.aeaweb.org/articles?id=10.1257/jel.20191450}{Abadie, A. (2021)}. Using synthetic controls:
Feasibility, data requirements, and methodological aspects.
\emph{Journal of Economic Literature}, 59(2), 391-425.}
}
}
\seealso{
\code{\link{scdata}}, \code{\link{scest}}, \code{\link{inference_sc}}
}
\author{
Matias Cattaneo, Princeton University. \email{cattaneo@princeton.edu}.

Yingjie Feng, Tsinghua University. \email{fengyj@sem.tsinghua.edu.cn}.

Filippo Palomba, Princeton University (maintainer). \email{fpalomba@princeton.edu}.

Rocio Titiunik, Princeton University. \email{titiunik@princeton.edu}.
}
